// frontend/src/types/api.d.ts
export interface paths {
// ...
    "/api/v1/top-up-orders": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["createTopUpOrder"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/top-up-orders/{orderId}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["getTopUpOrder"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/wallet": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["get"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        CreateRequest: {
            productCode: string;
        };
// ...
        OrderResponse: {
            /** Format: int32 */
            amount: number;
            /** Format: int32 */
            bonusShellAmount: number;
            /** Format: int32 */
            creditedShellAmount: number;
            currency: string;
            /** Format: uuid */
            orderId: string;
            orderName: string;
            /** Format: int32 */
            paidShellAmount: number;
            productCode: string;
            /** @enum {string} */
            status: "CREATED" | "PAYMENT_PENDING" | "PAID" | "CREDITED" | "FAILED" | "CANCELED" | "REFUNDED";
        };
// ...
        WalletResponse: {
            /** Format: int32 */
            balance: number;
            /** Format: int32 */
            bonusBalance: number;
            currency: string;
            /** Format: int32 */
            paidBalance: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}